fix: Dont attempt to remove uploaded tmp file if it's already removed

+++ packages/api/utils/upload.ts
@@ -119,7 +119,13 @@ export async function uploadAsset(
       fileName,
     };
   } finally {
-    if (tempFilePath) {
+    if (
+      tempFilePath &&
+      (await fs.promises
+        .access(tempFilePath)
+        .then(() => true)
+        .catch(() => false))
+    ) {
       await fs.promises.unlink(tempFilePath).catch(() => ({}));
     }
   }
